python-backend-conf/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.9.2
python-dotenv==1.0.0
supabase==2.3.0
groq==0.4.1
openai-agents==0.0.12
asyncpg==0.29.0
rapidfuzz==3.6.1